# docs/swagger.yaml
definitions:
  Models.Book:
    properties:
      author:
        type: string
      category:
        $ref: '#/definitions/Models.Category'
        description: ถ้าอยากให้เเสดงข้อมูลตอน Query ให้ใส่ Preload
      categoryID:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  Models.Category:
    properties:
      id:
        description: gorm.Model
        type: integer
      name:
        type: string
    type: object
  auth.Login:
    properties:
      jwt_password:
        type: string
      jwt_username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/category/:id:
    get:
      consumes:
      - application/json
      description: Get Category by id
      operationId: ListCategory
      parameters:
      - description: id of Category to be gotten
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Models.Category'
      security:
      - ApiKeyAuth: []
      summary: Get Category
      tags:
      - Category
  /book/:id:
    get:
      consumes:
      - application/json
      description: Get Book by id
      operationId: Book
      parameters:
      - description: id of Book to be gotten
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Models.Book'
      security:
      - ApiKeyAuth: []
      summary: Get Book
      tags:
      - Book
  /login:
    post:
      consumes:
      - application/json
      description: Get LoginHandler by id
      operationId: LoginHandler
      parameters:
      - description: id of LoginHandler to be gotten
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Login'
      security:
      - ApiKeyAuth: []
      summary: Post LoginHandler
      tags:
      - Auth
swagger: "2.0"